refactor generateEllipticKeys
